Pin platform to 6.11.0
Build issue on 6.12.0 (latest) that just started showing up:

```plaintext
fatal: not a git repository (or any of the parent directories): .git
```

This appears to be due to unbound dependency in
idf-component-manager, see this issue:

https://github.com/espressif/idf-component-manager/issues/97

Unfortunately platformio trails behind esp-idf releases, so we don't
have a simple workaround that I can figure out. Just pin to older
version for now.